examples/example_code-20260130T072212Z-3-001/example_code/requirements.txt
frappe-client>=0.0.3
python-dotenv>=0.19.0
requests>=2.27.0